requirements.txt
alembic==1.10.3
anyio==3.6.2
autopep8==2.0.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
fastapi==0.95.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
